+++ src/mods/JusticeForF7/thunderstore/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2026.717.0
+
+- Added target arrows to F7 hide mode.
+- Added character health bars to F7 hide mode.
+
 ## 2026.217.0
 
 - Fixed character names and zones being hidden on character selection screen